requirements.txt
tensorflow=2.1.0
flask
flask-cors
wget
numpy
pillow
keras